Adjunto de listas

+++ cheezery/app/src/main/java/ordunio/maribel/cheezery/Product.kt
@@ -1,6 +1,6 @@
 package ordunio.maribel.cheezery
 
-data class Product(var name: String, var image: Int, var desc: String, var price: Double) {
+data class Product(var name: String, var image: Int, var desc: String, var price: Int) {
 
 
 }
